+++ src/printer.ts
@@ -199,7 +199,23 @@ function handleBinaryishExpression(path: AstPath, print: printFn): Doc {
   //  .c() > d
   docs.push(group(leftDoc));
   docs.push(" ");
-  docs.push(groupConcat([operationDoc, line, rightDoc]));
+
+  // If the left child of a binaryish expression has an end of line comment,
+  // we want to make sure that comment is printed with the left child and
+  // followed by a hardline. Otherwise, it will lead to unstable comments in
+  // certain situation, because the EOL comment might become attached to the
+  // entire binaryish expression after the first format.
+  const leftChildHasEndOfLineComment =
+    node.left.comments?.filter(
+      (comment: AnnotatedComment) =>
+        comment.trailing && comment.placement === "endOfLine",
+    ).length > 0;
+
+  if (leftChildHasEndOfLineComment) {
+    docs.push(groupConcat([operationDoc, hardline, rightDoc]));
+  } else {
+    docs.push(groupConcat([operationDoc, line, rightDoc]));
+  }
   return groupConcat(docs);
 }